# docs/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bank REST API",
    "description": "Это API предоставляет эндпоинты для управления банковскими картами.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/api/admin/users/{userId}/roles/update": {
      "put": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Обновить роли пользователя",
        "description": "Полностью заменяет набор ролей пользователя на указанный в запросе.",
        "operationId": "updateRoles",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRolesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    },
...
    "/api/admin/users/{userId}/roles/remove": {
      "post": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Удалить роль у пользователя",
        "description": "Удаляет указанную роль у пользователя.",
        "operationId": "removeRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    },
    "/api/admin/users/{userId}/roles/add": {
      "post": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Добавить роль пользователю",
        "description": "Добавляет указанную роль пользователю.",
        "operationId": "addRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    },
    "/api/admin/users/create": {
      "post": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Создать пользователя",
        "description": "Администратор создаёт нового пользователя с указанными ролями.",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/cards/{id}/activate": {
      "post": {
        "tags": [
          "Admin Cards"
        ],
        "summary": "Активировать карту",
        "description": "Администратор активирует карту по ID карты.",
        "operationId": "activateCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    },
    "/api/admin/cards/{cardId}/block": {
      "post": {
        "tags": [
          "Admin Cards"
        ],
        "summary": "Заблокировать карту",
        "description": "Администратор блокирует карту по ID карты.",
        "operationId": "blockCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    },
    "/api/admin/cards/create": {
      "post": {
        "tags": [
          "Admin Cards"
        ],
        "summary": "Создать карту для пользователя",
        "operationId": "createCard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/transactions/{cardId}/get": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Получить транзакции по карте",
        "description": "Возвращает как входящие, так и исходящие транзакции",
        "operationId": "getTransactions",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/cards/{cardId}/balance": {
      "get": {
        "tags": [
          "User Cards"
        ],
        "summary": "Получить баланс карты",
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{id}/delete": {
      "delete": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Удалить пользователя",
        "description": "Удаляет пользователя по его ID.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    },
    "/api/admin/cards/{cardId}/delete": {
      "delete": {
        "tags": [
          "Admin Cards"
        ],
        "summary": "Удалить карту",
        "description": "Удаляет карту из системы по ID карты.",
        "operationId": "deleteCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешно"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UpdateUserRolesRequest": {
        "required": [
          "roles"
        ],
        "type": "object",
        "properties": {
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateUserRequest": {
        "required": [
          "email",
          "password",
          "roles",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateCardRequest": {
        "required": [
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "expiresInYears": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CardResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "ownerId": {
            "type": "integer",
            "format": "int64"
          },
          "maskedNumber": {
            "type": "string"
          },
          "decryptedNumber": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED",
              "PENDING_BLOCK"
            ]
          },
          "balance": {
            "type": "number"
          },
          "expiry": {
            "type": "string",
            "format": "date"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "TransactionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fromCardId": {
            "type": "integer",
            "format": "int64"
          },
          "toCardId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}